+++ demo-app/app/controllers/testing/drawers_controller.rb
@@ -1,0 +1,7 @@
+class Testing::DrawersController < ApplicationController
+  def index
+  end
+
+  def show
+  end
+end