+++ demo-app/config/initializers/ultimate_turbo_modal_tailwind.rb
@@ -1,1 +1,0 @@
-../../../lib/generators/ultimate_turbo_modal/templates/flavors/tailwind.rb