+++ lib/phlex/deferred_render_with_main_content.rb
@@ -4,7 +4,7 @@ module Phlex
   module DeferredRenderWithMainContent
     def view_template(&block)
       output = capture(&block)
-      super { respond_to?(:unsafe_raw) ? unsafe_raw(output) : raw(output) }
+      super { raw_html(output) }
     end
   end
 end