+++ lib/ultimate_turbo_modal/helpers/view_helper.rb
@@ -5,5 +5,10 @@ module UltimateTurboModal::Helpers
     def modal(**, &)
       render(UltimateTurboModal.new(request:, **), &)
     end
+
+    def drawer(position: nil, size: nil, **options, &block)
+      cfg = UltimateTurboModal.configuration.drawer_config
+      modal(_drawer_position: position || cfg.position, size: size, **options, &block)
+    end
   end
 end